feat: prep the identity module
