+++ .changeset/stale-waves-wink.md
@@ -2,4 +2,5 @@
 '@coinbase/onchainkit': minor
 ---
 
-- **feat**: added initial version of `getEASAttestations`, which helps getting the user attestations from the Ethereum Attetation Service (EAS).
+- **feat**: moved identity files into the `identity` folder. By @zizzamia
+- **feat**: added initial version of `getEASAttestations`, which helps getting the user attestations from the Ethereum Attetation Service (EAS). By @alvaroraminelli #126
