+++ site/docs/pages/identitykit/introduction.mdx
@@ -10,6 +10,24 @@ It provides a set of tools to resolve Ethereum addresses to ENS names, and more.
 
 To assist you in engaging with onchain Identity, here is the Identity Kit which includes:
 
+In order to use **Identity Kit** you must also install `graphql@14 graphql-request@6` in your app, alongside Onchainkit.
+
+:::code-group
+
+```bash [npm]
+npm install @coinbase/onchainkit graphql@14 graphql-request@6
+```
+
+```bash [yarn]
+yarn add @coinbase/onchainkit graphql@14 graphql-request@6
+```
+
+```bash [pnpm]
+pnpm add @coinbase/onchainkit graphql@14 graphql-request@6
+```
+
+:::
+
 - Components:
   - [`<Avatar />`](/identitykit/avatar): A component to display an ENS avatar.
   - [`<Name />`](/identitykit/name): A component to display an ENS name.
