+++ src/core/index.ts
@@ -1,5 +1,6 @@
 // 🌲☀️🌲
-export { getEASAttestations } from './getEASAttestations';
+// This file is the entry point for the core utilities
+// that have no dependencies on external libraries.
 export { getFrameHtmlResponse } from './getFrameHtmlResponse';
 export { getFrameMetadata } from './getFrameMetadata';
 export { getFrameMessage } from './getFrameMessage';
