+++ src/definitions/optimism.ts
@@ -1,5 +1,5 @@
 import { optimism } from 'viem/chains';
-import { EASChainDefinition } from '../core/types';
+import { EASChainDefinition } from '../identity/types';
 
 // More details in https://docs.optimism.io/chain/identity/schemas
 export const easChainOptimism: EASChainDefinition = {
