+++ src/identity/components/Avatar.test.tsx
