+++ src/identity/index.ts
@@ -1,0 +1,7 @@
+// 🌲☀️🌲
+export { Avatar } from './components/Avatar';
+export { Name } from './components/Name';
+export { useAvatar } from './hooks/useAvatar';
+export { useName } from './hooks/useName';
+export { getEASAttestations } from './getEASAttestations';
+export type { EASSchemaUid, EASAttestation, EASChainDefinition } from './types';
